+++ lib/routes/home/home_tab.dart
@@ -1,16 +1,20 @@
+import "package:flow/data/exchange_rates.dart";
 import "package:flow/data/transactions_filter.dart";
 import "package:flow/data/upcoming_transactions.dart";
 import "package:flow/entity/transaction.dart";
 import "package:flow/objectbox.dart";
 import "package:flow/objectbox/actions.dart";
 import "package:flow/prefs.dart";
+import "package:flow/services/exchange_rates.dart";
 import "package:flow/utils/optional.dart";
 import "package:flow/widgets/default_transaction_filter_head.dart";
 import "package:flow/widgets/general/wavy_divider.dart";
 import "package:flow/widgets/grouped_transaction_list.dart";
 import "package:flow/widgets/home/greetings_bar.dart";
+import "package:flow/widgets/home/home/flow_cards.dart";
 import "package:flow/widgets/home/home/no_transactions.dart";
 import "package:flow/widgets/home/transactions_date_header.dart";
+import "package:flow/widgets/rates_missing_warning.dart";
 import "package:flow/widgets/utils/time_and_range.dart";
 import "package:flutter/material.dart";
 import "package:moment_dart/moment_dart.dart";
@@ -94,6 +98,8 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
           ),
       builder: (context, snapshot) {
         final DateTime now = DateTime.now().startOfNextMinute();
+        final ExchangeRates? rates =
+            ExchangeRatesService().getPrimaryCurrencyRates();
         final List<Transaction>? transactions = snapshot.data;
 
         final Widget header = DefaultTransactionsFilterHead(
@@ -123,7 +129,8 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
                   ),
                 ),
               (_, true) => Expanded(
-                  child: buildGroupedList(context, now, transactions ?? []),
+                  child:
+                      buildGroupedList(context, now, transactions ?? [], rates),
                 ),
               (_, false) => const Expanded(
                   child: Center(
@@ -141,6 +148,7 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
     BuildContext context,
     DateTime now,
     List<Transaction> transactions,
+    ExchangeRates? rates,
   ) {
     final Map<TimeRange, List<Transaction>> grouped = transactions
         .where((transaction) => !transaction.transactionDate.isAfter(now))
@@ -153,6 +161,20 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
         currentFilter.accounts?.isNotEmpty != true;
 
     return GroupedTransactionList(
+      header: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const SizedBox(height: 12.0),
+          FlowCards(
+            transactions: transactions,
+            rates: rates,
+          ),
+          if (rates == null) ...[
+            const SizedBox(height: 12.0),
+            RatesMissingWarning(),
+          ],
+        ],
+      ),
       controller: widget.scrollController,
       transactions: grouped,
       futureTransactions: groupedFuture,